feat: let tidy provide behavior not defined by editorconfig
As of now, the presence of an '.editorconfig' file instructs tidy to not
run. Some users use editorconfig without instructing it to handle
whitespace.

In the case where the user wants to use editorconfig but let
tidy handle whitespace, a new option has been provided:
`provide_undefined_editorconfig_behavior`. This option is off by
default. When turned on, tidy will work when an editorconfig file is
found and when trimming whitespace is not defined. It will also remove
newlines at the end of the buffer without interfering with the
'insert_final_newline' option.