refactor: trim lines modified in insert mode, not in normal
In normal mode, actions like undo, redo, and deleting lines will affect
what is trimmed. This leads to trims on lines that were not really
touched by the user, giving errant trimmings.

Trim only the lines touched in insert mode. This means that lines that
the user edited in normal will not be trimmed (not ideal), but there
should be no errant trimmings (better than errors).